improve(sdk): remove index.d files, improve event types
