--- packages/sdk/src/index.d.ts
@@ -1,1 +1,0 @@
-export * from "./libs/index.d.ts";
