--- packages/sdk/src/libs/clients/registry/index.d.ts
@@ -1,3 +1,0 @@
-import * as Client from ".";
-export type Instance = ReturnType<typeof Client.connect>;
-
