--- packages/sdk/src/libs/tables/blocks/index.d.ts
@@ -1,6 +1,0 @@
-export type BlockType = {
-  id?: string;
-  number: number;
-  hash: string;
-  timestamp: number;
-};
