+++ packages/sdk/src/libs/tables/blocks/index.ts
@@ -1,1 +1,8 @@
 export { default as JsMap } from "./js-map";
+
+export type Block = {
+  id?: string;
+  number: number;
+  hash: string;
+  timestamp: number;
+};
