+++ packages/sdk/src/libs/tables/blocks/js-map.ts
@@ -1,13 +1,13 @@
 import Base from "../base";
 import Store from "../../stores/js-map";
-import type { BlockType } from "./index.d";
+import type { Block } from ".";
 
 const BlocksTable = () => {
-  function makeId(data: BlockType) {
+  function makeId(data: Block) {
     return data.number.toString();
   }
-  const store = Store<string, BlockType>();
-  const table = Base<string, BlockType>({ makeId, type: "Block" }, store);
+  const store = Store<string, Block>();
+  const table = Base<string, Block>({ makeId, type: "Block" }, store);
   return table;
 };
 export default BlocksTable;
